Fix unneeded read in `trace_parser`
The `parse_from_string` method in `trace_parser.cc` currently reads in one delta
for each active thread. However, `tracer_tool.cu` does not write the delta for
the first thread since it will always be zero. This commit updates the behavior
of `trace_parser` to match the intended trace format.

Even though this bug was present, it had no effect in the past since the
`base_delta_decompress` function is correct.