child process should ignore SIGINT
When using GDB to debug the simulator/post-processor, sending Ctrl-C to
the debugger will send a SIGINT to both the parent and child process.
The parent process pauses as expected, but without this patch, the child process will immediately terminate. This patch ensures the child process can persist while the simulator or post-processor is being debugged.