Add the function to generate SASS instruction set sequences with unique kernel_id and PC value
Add the function to generate SASS instruction set sequences with unique kernel_id and PC value, as someone may not want to see all instructions executed by all warps, but only the sequence of instructions with unique PC value whitin a single kernel. As long as the `DUMP_INTERSET` switch is turned on, the instruction set `instrset.csv` can be obtained in the `traces/` folder.